requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
executing==2.0.1
fastapi==0.110.0
google==3.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.4.1
google-resumable-media==2.7.0
google-search-results==2.4.2
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==2.10
install==1.3.5
ipython==8.23.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
notion-client==2.2.1
oauthlib==3.2.2
orjson==3.9.15
parso==0.8.3
pexpect==4.9.0
pillow==10.3.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.17.2
pyparsing==3.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.3
requests==2.25.1
requests-oauthlib==2.0.0
rsa==4.9
serpapi==0.1.5
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.36.3
tenacity==8.2.3
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.10.0
ujson==5.9.0
uritemplate==4.1.1
urllib3==1.26.4
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
